@jeremyevans 2025-01-16
Probably good to test here, but once tested, it should be submitted upstream to https://github.com/ruby/forwardable.

I'm guessing that this slowdown:

```
delegate with splat and double splat: 13665385.4 i/s - 2.53x  slower
delegate with splat with name: 13622624.9 i/s - 2.54x  slower
 delegate with splat: 13466096.0 i/s - 2.57x  slower
```

Is due to the array allocation for the positional argument splat (named or anonymous).  The anonymous keyword argument splat doesn't allocate because it uses nil.  Note that the anonymous positional splat allocation is avoided if there is a caller-side positional splat (e.g. `o.foo_with_splat(*array)`, with `array = []` defined outside the benchmark block).