@tenderlove 2025-01-16
> Probably good to test here, but once tested, it should be submitted upstream to https://github.com/ruby/forwardable.

Thanks @jeremyevans, I totally forgot this was probably in a gem. 🤦‍♂️