@tenderlove 2025-01-16
Actually I didn't realize forwadable is using ruby2_keywords. I don't really want to deal with this at the moment so I'm going to close this pr. 😭